requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0

# Audio and Speech Recognition
sounddevice>=0.4.5
numpy>=1.21.0
torch>=2.0.0
# Use PyPI version of whisper instead of Git version
openai-whisper>=20230918
pygame>=2.5.0  # For audio playback

# GUI
# tkinter is included with Python and not installable via pip

# Computer Control
pyautogui>=0.9.53  # For screenshots on Windows
python-xlib>=0.33  # For Linux system control
pyobjc-framework-Quartz>=9.0  # For macOS system control (optional) 